% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_a_song.R
\name{add_a_song}
\alias{add_a_song}
\title{Add a song to your existing database}
\usage{
add_a_song(database, song)
}
\arguments{
\item{database}{a dataframe containing music data}

\item{song}{a new song pulled from an API}
}
\value{
the dataframe with another row, containing the song
}
\description{
Add a song to your existing database
}
\examples{
\dontrun{
add_a_song(spotify_track_data,song)
}
}
