% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_time.R
\name{attributes_time}
\alias{attributes_time}
\title{Plot Attributes Over Time}
\usage{
attributes_time(df1, df1_title, df1_year_col, df2, df2_title, df2_year_col,
  attributes, boxplot, timerange, title_vector)
}
\arguments{
\item{df1}{a data.frame}

\item{df1_title}{title of your first df (ex: topsongs)}

\item{df1_year_col}{a number (column number of the "year" column)}

\item{df2}{a data.frame}

\item{df2_title}{title of your second df (ex: averagesongs)}

\item{df2_year_col}{a number (column number of the "year" column)}

\item{attributes}{a vector containing strings (names of columns to be plotted, ex: c("energy","danceability"))}

\item{boxplot}{a boolean (TRUE = plot as boxplot, FALSE = plot as averages)}

\item{timerange}{a 2x1 vector (years to be plotted over, ex: c(1990, 2010))}

\item{title_vector}{a string vector with df title names, defines what will be plotted (ex: c(df1_title) will plot only the first dataframe, c(df1_title, df2_title) will plot both)}
}
\value{
a plot
}
\description{
This function plots the features of 2 different datasets (in dataframes) over time in the same graph.
The dataframes should have overlapping features and contain a column "year" over which these features will be plotted.
An example of 2 datasets are: top songs & average songs with "danceability" and "energy" attributes over years 1990 to 2010.
}
\examples{
\dontrun{
attributes_time(topsongs, "Top Songs", 1, averagesongs, "Average Songs", 4,
c("energy","speechiness"), FALSE, c(1960,2015), c("Top Songs","Average Songs"))
}
}
