% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averagesongs.R
\docType{data}
\name{averagesongs}
\alias{averagesongs}
\title{Average songs}
\format{A data frame with 150425 rows and 23 variables:
\describe{
  \item{track_name}{name of the track}
  \item{track_uri}{uri of the track}
  \item{album_name}{name of the album}
  \item{year}{year of the album}
  \item{danceability}{feature between 0-1}
  \item{energy}{feature between 0-1}
  \item{key}{feature between 0-1}
  \item{loudness}{feature between 0-1}
  \item{mode}{feature between 0-1}
  \item{speechiness}{feature between 0-1}
  \item{acousticness}{feature between 0-1}
  \item{instrumentalness}{feature between 0-1}
  \item{liveness}{name of the artist}
  \item{valence}{name of the artist}
  \item{tempo}{feature between 0-1}
  \item{type}{feature between 0-1}
  \item{uri}{name of the artist}
  \item{track_href}{name of the artist}
  \item{analysis_url}{feature between 0-1}
  \item{duration_ms}{feature between 0-1}
  \item{time_signature}{feature between 0-1}
  \item{artist_name}{name of the artist}
  \item{artist_id}{id of the artist}
}}
\usage{
averagesongs
}
\description{
A dataset containing the average songs from 1960-2018
}
\keyword{datasets}
