% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_data.R
\docType{data}
\name{bb_data}
\alias{bb_data}
\title{bb_data}
\format{A data frame with 5498 rows}
\usage{
bb_data
}
\description{
A dataset containing the Billboard songs for each year with clusters and principal components included
}
\keyword{datasets}
