% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard_music_dataframe.R
\docType{data}
\name{billboard_music_dataframe}
\alias{billboard_music_dataframe}
\title{Billboard Music Dataframe}
\format{A data frame with 5497 rows and 23 variables. Each song has several attributes (from Spotify API)}
\usage{
billboard_music_dataframe
}
\description{
A dataset containing the top billboard charts from 1960-2018
}
\keyword{datasets}
