% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_songs_k.R
\name{classify_songs_k}
\alias{classify_songs_k}
\title{Add a column for songs containing k means cluster for a given year}
\usage{
classify_songs_k(song, year, k_model)
}
\arguments{
\item{song}{a new song pulled from an API}

\item{year}{the year to be compared with}

\item{k_model}{the model for k means clustering}
}
\value{
the dataframe with additional column for cluster
}
\description{
Add a column for songs containing k means cluster for a given year
}
\examples{
\dontrun{
classify_songs_k(song,year,k_model)
}
}
