% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_data.R
\docType{data}
\name{dim_pc_2}
\alias{dim_pc_2}
\title{dim_pc_2}
\format{A yearly list of linear models}
\usage{
dim_pc_2
}
\description{
A list of linear models for predicting the second principal component using song features
}
\keyword{datasets}
