% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_new_songs.R
\name{format_new_songs}
\alias{format_new_songs}
\title{Format the songs that the user inputs into the same format as the billboard top charts}
\usage{
format_new_songs(songs)
}
\arguments{
\item{songs}{a new song pulled from an API}
}
\value{
the dataframe with another row, containing the song
}
\description{
Format the songs that the user inputs into the same format as the billboard top charts
}
\examples{
\dontrun{
format_new_songs(songs)
}
}
