% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_new_songs_logit.R
\name{format_new_songs_logit}
\alias{format_new_songs_logit}
\title{Format New Songs Logit
Is a function that prepares musical data from spotify to be fit into the prediction function of this App}
\usage{
format_new_songs_logit(songs)
}
\arguments{
\item{songs}{a dataframe containing music information}
}
\value{
a dataframe containing music information in a different format
}
\description{
Format New Songs Logit
Is a function that prepares musical data from spotify to be fit into the prediction function of this App
}
\examples{
\dontrun{
format_new_songs_logit(tracks_from_spotify)
}
}
