% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probability_of_billboard.R
\name{get_probability_of_billboard}
\alias{get_probability_of_billboard}
\title{Get Probability of Song being on Billboard}
\usage{
get_probability_of_billboard(input_song, model)
}
\arguments{
\item{input_song}{a named list of all attributes for a song to analyze its probability to be on the billboard charts
the input song has the following attributes:
"danceability"
"energy"
"key"
"loudness"
"mode"
"speechiness"
"acousticness"
"instrumentalness"
"liveness"
"valence"
"tempo"
"duration_ms"
"artist_name"
"track_name"
"album_year"}

\item{model}{The list of logistic regression models that predict probability that a song will be on the billboard charts}
}
\value{
a dataframe containing the probability a song will be on the billboard charts throughout time
}
\description{
This function gets the probability that a song will be on the billboard charts throughout time
}
\examples{
\dontrun{
get_probability_of_billboard(input_song)

}
}
