% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_and_features.R
\name{get_tracks_and_features}
\alias{get_tracks_and_features}
\title{Get Tracks and Features}
\usage{
get_tracks_and_features(search_string,
  access_token = get_spotify_access_token())
}
\arguments{
\item{search_string}{a string containing a name of the artist or song}

\item{access_token}{the access token for to the Spotify API}
}
\value{
a 23 x 20 dataframe containing the first 20 matches from spotify including music characteristics such as energy, danceability etc.
}
\description{
get_tracks_and_features returns the 20 first matches of a string query to the spotify API.
The returned dataframe contains all the characteristics needed to make use of the musictasteR plots
}
\examples{
get_tracks_and_features("thriller")
}
