% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_artists.R
\name{get_tracks_artists}
\alias{get_tracks_artists}
\title{Get Tracks Artists}
\usage{
get_tracks_artists(track_artist_name,
  access_token = get_spotify_access_token())
}
\arguments{
\item{track_artist_name}{A string with track name or an artist name}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
a dataframe containing the track information
}
\description{
This functions takes a string and returns a data frame with track information for 20 songs
from Spotify's search endpoin
}
\examples{
\dontrun{
#### Get track information for the top 20 Spotify matches for "Thriller"
thriller <- get_tracks_artists(track_artist_name = "Thriller")
}
}
