% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hover.plot}
\alias{hover.plot}
\title{Hover Plot}
\usage{
hover.plot(database, filtered_year, year_col, x_axis, y_axis,
  track_name = track_name, artist_name = artist_name,
  title = "Billboard Top 100 musical charts of ")
}
\arguments{
\item{database}{the database containing musical data you want to analyze. We suggest the billboard::spotify_track_data}

\item{filtered_year}{a string, the year you want to look at}

\item{year_col}{a column of your df containing the years}

\item{x_axis}{a column of your df, the first dimension you want to plot against - such as energy}

\item{y_axis}{a column of your df, the second dimension you want to plot against - such as danceability}

\item{track_name}{a column of your df, information 1 you want to displayed when hovering - such as track_name}

\item{artist_name}{a column of your df, information 2 you want to displayed when hovering - such as artist_name}

\item{title}{a string, the title of your plot}
}
\value{
a plot
}
\description{
Hover Plot
}
\examples{
\dontrun{
hover.plot(spotify_track_data,"1999",year,energy,danceability,track_name,artist_name,"Charts")
}
}
