% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hover.plot.shiny}
\alias{hover.plot.shiny}
\title{Hover Plot Shiny}
\usage{
hover.plot.shiny(data, x, y, chosen_year)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the x variable to be displayed in the plot}

\item{y}{the y variable to be displayed in the plot}

\item{chosen_year}{the year for which we}
}
\value{
a plotly plot
}
\description{
specific function for the shiny app
}
\examples{
\dontrun{
hover.plot.shiny(spotify_track_data, input$x,input$y,input$year)
}
}
