% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover_plot_shiny.R
\name{hover_plot_shiny}
\alias{hover_plot_shiny}
\title{Hover Plot Shiny}
\usage{
hover_plot_shiny(new_music, x, y, chosen_year)
}
\arguments{
\item{new_music}{a dataframe}

\item{x}{a column name of the billboard data frame the first characteristic the music should be plotted against}

\item{y}{a column name of the billboard data frame the second characteristic the music should be plotted against}

\item{chosen_year}{the year for which the songs are shown, between 1960 and 2015 if using the billboard data}
}
\value{
a plotly plot showing the top billboard charts filtered by year, plotted against x and y, and the songs added by the user
}
\description{
Hover Plot Shiny is a function that allows you to plot your favorite song against the top billboard charts from 1960 - 2015.
}
\examples{
\dontrun{
hover_plot_shiny(new_music,energy,danceability,"1994")
}
}
