% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_data.R
\docType{data}
\name{k_model}
\alias{k_model}
\title{k_model}
\format{A yearly list of clustering models}
\usage{
k_model
}
\description{
A list of clustering models for predicting the cluster of each song
}
\keyword{datasets}
