% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_list.R
\docType{data}
\name{log_model_list}
\alias{log_model_list}
\title{Logistic Regression Models}
\format{A list containing a logistic regression model per year frmo 1960 - 2015}
\usage{
log_model_list
}
\description{
Logistic Regression Models
}
\keyword{datasets}
