% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_avg.R
\name{plot_time_avg}
\alias{plot_time_avg}
\title{Title}
\usage{
plot_time_avg(df, vars, boxplot, timerange)
}
\arguments{
\item{df}{a dataframe containing one column "year"}

\item{vars}{attributes to be plotted}

\item{boxplot}{a boolean (T/F)}

\item{timerange}{a vector of integers (years to be plotted over)}
}
\value{
a plot
}
\description{
Title
}
\examples{
plot_time_avg(df, c("danceability","valence"),FALSE,c(2010,2015))
}
