% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pc_lm.R
\name{predict_pc_lm}
\alias{predict_pc_lm}
\title{Add columns for songs containing principal dimensions based on pre-created linear models}
\usage{
predict_pc_lm(song, year, mod1, mod2)
}
\arguments{
\item{song}{a new song pulled from an API}

\item{year}{the year to be compared with}

\item{mod1}{the model for dimension 1 of PCA}

\item{mod2}{the model for dimension 2 of PCA}
}
\value{
the dataframe additional columns for dimensions 1 and 2 ie dim_1,dim_2
}
\description{
Add columns for songs containing principal dimensions based on pre-created linear models
}
\examples{
\dontrun{
predict_pc_lm(song,year,mod1,mod2)
}
}
