% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubasic-package.R
\docType{package}
\name{rubasic-package}
\alias{rubasic}
\alias{rubasic-package}
\title{rubasic: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
