% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_test_pull.R
\docType{data}
\name{spotify_test_pull}
\alias{spotify_test_pull}
\title{Spotify Test Pull}
\format{A data frame with 20 rows and 9 variables}
\usage{
spotify_test_pull
}
\description{
A dataset containing the first 20 results when calling the Spotify API with the string "Golddigger"
}
\keyword{datasets}
