% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_test_pull_formated_1.R
\docType{data}
\name{spotify_test_pull_formated_1}
\alias{spotify_test_pull_formated_1}
\title{spotify_test_pull_formated_1}
\format{a 23 x 20 dataframe containing the first 20 matches from spotify including music characteristics such as energy, danceability etc.}
\usage{
spotify_test_pull_formated_1
}
\description{
A preformatted data set based on an API call that is used by hover_plot_shiny and plot_songs_clusters
}
\keyword{datasets}
