% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_rest_pull_with_characteristics.R
\docType{data}
\name{spotify_test_pull_with_characteristics}
\alias{spotify_test_pull_with_characteristics}
\title{Spotify Test Pull With Characteristics}
\format{a 23 x 20 dataframe containing the first 20 matches from spotify including music characteristics such as energy, danceability etc.}
\usage{
spotify_test_pull_with_characteristics
}
\description{
A dataset containing the first 20 results when calling the Spotify API with the string "Golddigger"
}
\keyword{datasets}
