% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probabilities.R
\name{plot_probabilities}
\alias{plot_probabilities}
\title{Title}
\usage{
plot_probabilities(input_dataframe, year_int_col_index, prob_col_index,
  track_name_col_index, true_song_year_index)
}
\arguments{
\item{input_dataframe}{the input song has the following attributes: "trackname", "year_int", "prob", "true_song_year"}

\item{year_int_col_index}{index for year_int col in the dataframe}

\item{prob_col_index}{index for prob col in the dataframe}

\item{track_name_col_index}{index for track_name col in the dataframe}

\item{true_song_year_index}{index for true song year boolean column}
}
\value{
return a ggplot
}
\description{
Title
}
\examples{
\dontrun{
plot_probabilities(input_dataframe)
}

}
