% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_songs_clusters.R
\name{plot_songs_clusters}
\alias{plot_songs_clusters}
\title{Plot Songs Cluster}
\usage{
plot_songs_clusters(songs, year_taken)
}
\arguments{
\item{songs}{a new song pulled from an API}

\item{year_taken}{the year to be compared with}
}
\value{
the dataframe with additional column for cluster
}
\description{
Return a plot element for a song and a year, in comparison to other billboard tracks in that year
}
\examples{
\dontrun{
plot_songs_clusters(songs,year_taken)
}
}
